# docs/openapi/novel-platform.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Novel Platform API",
    "version": "0.1.0",
    "description": "Versioned REST contract for the novel platform backend. Browser clients use the same-origin Next BFF: replace the backend prefix `/api/v1` with `/api/novel` for the `public`, `account`, `author`, and `admin` route groups. The BFF owns browser cookies, CSRF validation, and the upstream internal headers; do not call protected backend routes from browser JavaScript directly. `/api/v1/auth/*` is an internal BFF-to-backend contract and is intentionally not exposed by the catch-all browser proxy. This contract describes implemented local platform behavior only. Phone/email verification, WeChat/QQ OAuth, external payment, and production Qwen credentials are not represented as available API integrations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Backend service. Protected and auth routes require BFF-only headers."
    }
  ],
...
  "paths": {
...
    "/api/v1/public/books/{bookId}/comments": {
      "get": {
        "tags": ["Public discovery"],
        "operationId": "getVisibleComments",
        "summary": "Page visible comments for one published book",
        "x-access": "public",
        "security": [],
        "x-response-data-schema": "#/components/schemas/CommentPageResponse",
        "parameters": [
          { "$ref": "#/components/parameters/PathBookId" },
          { "$ref": "#/components/parameters/ChapterIdQuery" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/InteractionPageSize" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/public/books/{bookId}/interactions": {
      "get": {
        "tags": ["Public discovery"],
        "operationId": "getPublicInteractionStats",
        "summary": "Get visible interaction aggregates for one book",
        "x-access": "public",
        "security": [],
        "x-response-data-schema": "#/components/schemas/InteractionStats",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/public/books/{bookId}/chapters/{chapterId}/annotations": {
      "get": {
        "tags": ["Public discovery"],
        "operationId": "getVisibleParagraphAnnotations",
        "summary": "Page publicly approved paragraph annotations",
        "x-access": "public",
        "security": [],
        "x-response-data-schema": "#/components/schemas/PublicAnnotationPageResponse",
        "parameters": [
          { "$ref": "#/components/parameters/PathBookId" },
          { "$ref": "#/components/parameters/PathChapterId" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/InteractionPageSize" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "novel_session",
        "description": "Opaque browser BFF session cookie. The browser never receives the backend session id."
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-CSRF",
        "description": "Required by the BFF on unsafe browser requests and must match its CSRF cookie/session binding."
      },
      "InternalBffKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Internal-Key",
        "description": "Private BFF-to-backend credential. Never send this header from a browser."
      },
      "BackendBffSession": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Bff-Session",
        "description": "Server-side backend session credential held by the BFF only. Never expose it to browser code."
      }
    },
    "parameters": {
...
      "PathBookId": { "name": "bookId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "PathChapterId": { "name": "chapterId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "ChapterIdQuery": { "name": "chapterId", "in": "query", "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0, "maximum": 100000 } },
      "InteractionPageSize": { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1 }, "description": "The controller accepts an interaction page size; callers should use a bounded value." },
...
    },
...
    "responses": {
      "Success": { "description": "Successful platform envelope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiEnvelope" } } } },
...
      "NotFound": { "description": "Requested resource does not exist or is not visible to the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer", "example": 200 }, "msg": { "type": "string", "example": "ok" }, "data": {} }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer" }, "msg": { "type": "string" }, "data": { "type": "null" } }
      },
...
      "Comment": { "type": "object", "required": ["id", "bookId", "userId", "authorName", "content", "status", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": ["integer", "null"], "format": "int64" }, "userId": { "type": "integer", "format": "int64" }, "authorName": { "type": "string" }, "content": { "type": "string" }, "status": { "type": "string", "enum": ["PENDING_REVIEW", "VISIBLE", "REJECTED"] }, "createdAt": { "type": "string", "format": "date-time" } } },
      "CommentPageResponse": { "type": "object", "required": ["items", "meta"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } }, "meta": { "$ref": "#/components/schemas/PageMeta" } } },
...
      "PublicParagraphAnnotation": { "type": "object", "required": ["id", "bookId", "chapterId", "authorName", "paragraphIndex", "selectionStart", "selectionEnd", "selectedText", "note", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": "integer", "format": "int64" }, "authorName": { "type": "string" }, "paragraphIndex": { "type": "integer" }, "selectionStart": { "type": "integer" }, "selectionEnd": { "type": "integer" }, "selectedText": { "type": "string" }, "note": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "PublicAnnotationPageResponse": { "type": "object", "required": ["items", "meta"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/PublicParagraphAnnotation" } }, "meta": { "$ref": "#/components/schemas/PageMeta" } } },
      "PageMeta": { "type": "object", "required": ["total", "page", "size"], "properties": { "total": { "type": "integer", "format": "int64", "minimum": 0 }, "page": { "type": "integer", "minimum": 0 }, "size": { "type": "integer", "minimum": 1 } } },
      "InteractionStats": { "type": "object", "required": ["visibleCommentCount", "ratingCount", "averageRating", "recommendationVoteCount", "monthlyVoteCount"], "properties": { "visibleCommentCount": { "type": "integer", "format": "int64" }, "ratingCount": { "type": "integer", "format": "int64" }, "averageRating": { "type": "number", "format": "double" }, "recommendationVoteCount": { "type": "integer", "format": "int64" }, "monthlyVoteCount": { "type": "integer", "format": "int64" } } },
...
    }
  }
}